Добавлены меню звука и управления
