+++ Assets/Scripts/Menu/GameMenu.cs
@@ -1,0 +1,14 @@
+
+
+namespace Menu
+{
+    public class GameMenu : MenuSection
+    {
+        public static bool ToggleButtonsEnabled { get; private set; }
+
+        public void SetToggleButtonBehavior(bool useToggle)
+        {
+            ToggleButtonsEnabled = useToggle;
+        }
+    }
+}
