+++ Assets/Scripts/Menu/MainMenu.cs
@@ -1,16 +1,9 @@
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 namespace Menu
 {
     public class MainMenu : MenuSection
-    
     {
-        public void PlayGame()
-        {
-            SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
-        }
-
         public void Exit()
         {
             Application.Quit();
