--- Assets/Scripts/Menu/OptionsMenu.cs
@@ -1,20 +1,0 @@
-﻿using UnityEngine;
-
-namespace Menu
-{
-    public class OptionsMenu: MenuSection
-    {
-        public static float VolumeLevel { get; private set; }
-        public static bool ToggleButtonsEnabled { get; private set; }
-        
-        public void SetVolume(float volume)
-        {
-            VolumeLevel = volume;
-        }
-
-        public void SetToggleButtonBehavior(bool useToggle)
-        {
-            ToggleButtonsEnabled = useToggle;
-        }
-    }
-}
