--- Assets/Scripts/Menu/OptionsMenu.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2bf1a1551b21440e82819bdef3f06928
-timeCreated: 1679502651
