+++ Assets/Scripts/Menu/SoundMenu.cs
@@ -1,0 +1,13 @@
+
+namespace Menu
+{
+    public class SoundMenu : MenuSection
+    {
+        public static float VolumeLevel { get; private set; }
+        
+        public void SetVolume(float volume)
+        {
+            VolumeLevel = volume;
+        }
+    }
+}
